// src/lib/api/model/strikeDtoLevel.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Wildcat Radio v2 API
 * Control plane — see docs/final-build-plan/03-API-CONTRACT.md
 * OpenAPI spec version: 0.0.1
 */

export type StrikeDtoLevel = typeof StrikeDtoLevel[keyof typeof StrikeDtoLevel];


export const StrikeDtoLevel = {
  NUMBER_1: 1,
  NUMBER_2: 2,
  NUMBER_3: 3,
} as const;